# backend/docs/swagger.yaml
basePath: /
definitions:
  auth.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      code:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  auth.PasswordResetRequest:
    properties:
      email:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  auth.TokenResponse:
    properties:
      token:
        type: string
    type: object
  auth.UpdatePasswordRequest:
    properties:
      email:
        type: string
      newPassword:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: alisfrk2004@gmail.com
    name: Дарья
  description: Серверная часть веб-приложения для коммуникации.
  termsOfService: http://swagger.io/terms/
  title: Buzz API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Авторизирует пользователя в приложении и выдает JWT-token
      parameters:
      - description: Данные для авторизации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь авторизирован
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Некорректный формат данных
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Неверные учётные данные
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/password-reset:
    post:
      consumes:
      - application/json
      description: Отправляет ссылку-подтверждение для восстановления пароля
      parameters:
      - description: Email пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Ссылка-подверждение отправлена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный email
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Запрос на восстановление пароля
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя и отправляет письмо с подтверждением
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
        "400":
          description: Некорректный формат данных
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "409":
          description: Неверные учётные данные
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
  /auth/update-password:
    post:
      consumes:
      - application/json
      description: Меняет пароль пользователя на новый
      parameters:
      - description: Токен и новый пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пароль успешно изменён
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный пароль
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Недействительный или просроченный токен
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Установка нового пароля
      tags:
      - auth
...
  /rooms/{id}/board:
    get:
      description: Возвращает сохранённое состояние доски для указанной комнаты. Если
        доска ещё не инициализирована, возвращает пустой объект {}.
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Текущее состояние доски
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователь не является участником комнаты
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить текущее состояние виртуальной доски
      tags:
      - board
...
  /ws/board:
    get:
      description: Устанавливает WebSocket-соединение для обмена обновлениями доски
        в реальном времени. Соединение должно содержать query-параметр roomId и заголовок
        Authorization с Bearer-токеном.
      parameters:
      - description: ID комнаты
        in: query
        name: roomId
        required: true
        type: string
      responses:
        "101":
          description: WebSocket-соединение установлено
        "400":
          description: Неверный id комнаты
        "401":
          description: Неавторизован
        "403":
          description: Пользователь не является участником комнаты
      security:
      - BearerAuth: []
      summary: Подключение к WebSocket для доски
      tags:
      - board
...
securityDefinitions:
  BearerAuth:
    description: Введите токен в формате "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"